expose findCellRectClosestToPos table helper
