+++ __tests__/helpers.js
@@ -1,20 +1,6 @@
-import {
-  createEditor,
-  doc,
-  p,
-  table,
-  tr as row,
-  td,
-  tdEmpty,
-  strong,
-  atomInline
-} from '../test-helpers';
+import { createEditor, doc, p, strong, atomInline } from '../test-helpers';
 import { Fragment } from 'prosemirror-model';
-import {
-  canInsert,
-  removeNodeAtPos,
-  findCellClosestToPos
-} from '../src/helpers';
+import { canInsert, removeNodeAtPos } from '../src/helpers';
 
 describe('helpers', () => {
   describe('canInsert', () => {
@@ -82,29 +68,6 @@ describe('helpers', () => {
       const newTr = removeNodeAtPos(4)(tr);
       expect(newTr).not.toBe(tr);
       expect(newTr.doc).toEqualDocument(doc(p('one')));
-    });
-  });
-
-  describe('findCellClosestToPos', () => {
-    it('should return a cell object closest to a given `$pos`', () => {
-      const {
-        state: { schema, tr }
-      } = createEditor(
-        doc(
-          table(row(td(p('one one')), tdEmpty), row(td(p('two two')), tdEmpty))
-        )
-      );
-      const cell = findCellClosestToPos(tr.doc.resolve(4));
-      expect(cell.node.type.name).toEqual('table_cell');
-      expect(cell.pos).toEqual(2);
-    });
-
-    it('should return `undefined` if there is no cell close to a given `$pos`', () => {
-      const {
-        state: { tr }
-      } = createEditor(doc(p('one')));
-      const cell = findCellClosestToPos(tr.doc.resolve(4));
-      expect(cell).toBeUndefined();
     });
   });
 });
