+++ src/index.js
@@ -2,4 +2,4 @@ export * from './selection';
 export * from './node';
 export * from './table';
 export * from './transforms';
-export { isNodeSelection, canInsert, findCellClosestToPos } from './helpers';
+export { isNodeSelection, canInsert } from './helpers';
